Add support to Allowed Groups
The list of LDAP group DNs that users must be members of to be granted access.
If a user is in any one of the listed groups, then that user is granted access.

If setted to an empty list will allow all users that have an LDAP account to
log in, without performing any group membership checks.

Fixes #3